+++ packages/alfa-rules/src/common/act/question.ts
@@ -248,6 +248,15 @@ export namespace Question {
       type: "boolean",
       message: `Is this element visible when it's focused?`,
     },
+    // R101 (experimental)
+    "main-landmark-elements": {
+      type: "node[]",
+      message: "Where are the main landmark elements of the page?",
+    },
+    "has-repeated-content-before-main": {
+      type: "boolean",
+      message: "Does the page have repeated content before its main content?",
+    },
     // R109 (experimental)
     "document-language": {
       type: "string",