+++ packages/alfa-rules/src/sia-dr83/rule.ts
@@ -49,6 +49,10 @@ const {
   isVisible,
 } = Style;
 
+/**
+ * @deprecated
+ * This rule has been deprecated due to a too-strict interpretation of WCAG.
+ */
 export default Rule.Atomic.of<Page, Text>({
   uri: "https://alfa.siteimprove.com/rules/sia-r83",
   requirements: [Criterion.of("1.4.4")],