+++ packages/alfa-graph/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @siteimprove/alfa-graph
 
+## 0.115.0
+
 ## 0.114.3
 
 ### Patch Changes